pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["FuckingSherlock"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2.8"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
